chore: remove outdated social bar
