+++ components/molecules/index.ts
@@ -2,3 +2,4 @@
 // Simple combinations of atoms functioning together
 
 export * from './AuthActions';
+export * from './SocialLink';
