+++ components/organisms/index.ts
@@ -2,3 +2,4 @@
 // Complex combinations of molecules and atoms
 
 export * from './Header';
+export * from './SocialBar';
