+++ tests/unit/SocialBar.test.tsx
@@ -1,6 +1,6 @@
 import { describe, it, expect, afterEach, vi } from 'vitest';
 import { render, screen, cleanup } from '@testing-library/react';
-import { SocialBar } from '@/components/profile/SocialBar';
+import { SocialBar } from '@/components/organisms/SocialBar';
 import type { SocialLink } from '@/types/db';
 
 // Mock fetch for API calls
@@ -40,11 +40,11 @@ describe('SocialBar', () => {
       />
     );
 
-    const socialButtons = screen.getAllByRole('button');
-    expect(socialButtons).toHaveLength(2);
+    const socialLinksElements = screen.getAllByRole('link');
+    expect(socialLinksElements).toHaveLength(2);
 
-    socialButtons.forEach((button) => {
-      expect(button).toHaveClass('cursor-pointer');
+    socialLinksElements.forEach((link) => {
+      expect(link).toHaveClass('cursor-pointer');
     });
   });
 
@@ -70,14 +70,14 @@ describe('SocialBar', () => {
       />
     );
 
-    const buttons = screen.getAllByRole('button');
-    expect(buttons[0]).toHaveAttribute('title', 'Follow on instagram');
-    expect(buttons[0]).toHaveAttribute(
+    const links = screen.getAllByRole('link');
+    expect(links[0]).toHaveAttribute('title', 'Follow on instagram');
+    expect(links[0]).toHaveAttribute(
       'aria-label',
       'Follow Test Artist on instagram'
     );
-    expect(buttons[1]).toHaveAttribute('title', 'Follow on twitter');
-    expect(buttons[1]).toHaveAttribute(
+    expect(links[1]).toHaveAttribute('title', 'Follow on twitter');
+    expect(links[1]).toHaveAttribute(
       'aria-label',
       'Follow Test Artist on twitter'
     );
